+++ silhouette/app/com/mohiva/play/silhouette/impl/services/GravatarService.scala
@@ -19,6 +19,7 @@
  */
 package com.mohiva.play.silhouette.impl.services
 
+import java.net.URLEncoder._
 import java.security.MessageDigest
 import javax.inject.Inject
 
@@ -33,9 +34,10 @@ import scala.concurrent.Future
  * Retrieves avatar URLs from the Gravatar service.
  *
  * @param httpLayer The HTTP layer implementation.
+ * @param settings The Gravatar service settings.
  */
-class GravatarService @Inject() (httpLayer: HTTPLayer) extends AvatarService
-  with Logger with ExecutionContextProvider {
+class GravatarService @Inject() (httpLayer: HTTPLayer, settings: GravatarServiceSettings = GravatarServiceSettings())
+  extends AvatarService with Logger with ExecutionContextProvider {
 
   /**
    * The execution context to handle the asynchronous operations.
@@ -51,7 +53,8 @@ class GravatarService @Inject() (httpLayer: HTTPLayer) extends AvatarService
   override def retrieveURL(email: String): Future[Option[String]] = {
     hash(email) match {
       case Some(hash) =>
-        val url = URL.format(hash)
+        val encodedParams = settings.params.map { p => encode(p._1, "UTF-8") + "=" + encode(p._2, "UTF-8") }
+        val url = (if (settings.secure) SecureURL else InsecureURL).format(hash, encodedParams.mkString("?", "&", ""))
         httpLayer.url(url).get().map(_.status match {
           case 200 => Some(url)
           case code =>
@@ -86,6 +89,18 @@ class GravatarService @Inject() (httpLayer: HTTPLayer) extends AvatarService
  * The companion object.
  */
 object GravatarService {
-  val URL = "http://www.gravatar.com/avatar/%s?d=404"
+  val InsecureURL = "http://www.gravatar.com/avatar/%s%s"
+  val SecureURL = "https://secure.gravatar.com/avatar/%s%s"
   val MD5 = "MD5"
 }
+
+/**
+ * The gravatar service settings object.
+ *
+ * @param secure Indicates if the secure or insecure URL should be used to query the avatar images. Defaults to secure.
+ * @param params A list of params to append to the URL.
+ * @see https://en.gravatar.com/site/implement/images/
+ */
+case class GravatarServiceSettings(
+  secure: Boolean = true,
+  params: Map[String, String] = Map("d" -> "404"))