+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 3.0 (2015-06-08)
+## 3.0 (2015-06-11)
 
 - Update to Play 2.4
 - Stateless and non-stateless CookieAuthenticator
+- Allow to customize the Gravatar service
+- Use scala.concurrent.duration.Duration instead of Int for duration based values
 - A lot of API enhancements
 
 ## 2.0 (2015-03-28)